평점관련 이슈 발생 및 0523 내에 해결할 예정 model을 만들거나 평점 삭제 및 수정을 어떻게 모델없이 할 수 있을지 고민해봐야할 듯

+++ server/movies/models.py
@@ -10,7 +10,7 @@ class Movie(models.Model):
     pubDate = models.CharField(max_length=256)
     director = models.CharField(max_length=256)
     actor = models.CharField(max_length=256)
-    userRating = models.FloatField()
+    userRating = models.IntegerField()
 
 
 class Review(models.Model):
